requirements.txt
instagrapi==2.0.0
pandas==2.1.3
openpyxl==3.1.2
matplotlib==3.8.2
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0